Make `GFp_BN_new()` test-only.
