+++ mk/ring.mk
@@ -127,6 +127,7 @@ RING_TEST_SRCS = $(addprefix $(RING_PREFIX), \
   crypto/constant_time_test.c \
   crypto/test/bn_test_convert.c \
   crypto/test/bn_test_lib.c \
+  crypto/test/bn_test_new.c \
   crypto/test/file_test.cc \
   $(NULL))
